describe project jupik
There is a course Java from Zero to Junior Plus on the learning platform
Stepik (<https://stepik.org>). The completion of the course is verified
by certificate. Beside it is still valuable to confirm the completion
with a text that demonstrates materials studied, a so called
*Portfelia*.

The text provided in this project is typed by me to solve or to prepare
to solve the issues of the course. But it is not intended to be a
precise solution of the only issues of the course.

This project can be used in accordance with rules and limitation of
License [GPLv3+][2]: GNU GPL version 3 or later
<https://gnu.org/licenses/gpl.html>. There is NO WARRANTY except
permitted by law.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>