[01-INTRO] describe the issue
In the first topic it is stated to be ready to be hired for a position
of a Junior Plus Java Programmer it is required to complete two parts
of the course. The part Level 1 and than the part Level 2.

To start with this it is requested to ask Google how to install a
proprietary text processor onto a proprietary OS. And them how to click
some pictograms to generate a project, to generate a Java class, to add
a start point of the program and to add logic to print a message onto a
terminal.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>