[03-OOP] add class of cats
The inheritance means that some properties of a class could be provided from
some generic class.

There are properties generic for all members of Felidae: the ability to eat
humans, the number of eyes and legs.

In further commits the class Cat will be used to solve a gastronomic issue.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>