[03-OOP] provide super class of figures
Provide the super class of figures with two properties, the length of
sides and a constactor to configure this properties on initialisation.

Indeed, the agragation MUST BE prefered over inheritance but it is
REQUIRED to provide a class not an interface.

In further commits this class will be used to solve a geometrics issue.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>