[04-ADVANCE] solve randomisation problem
The operation toString represents an object as a string. The default
implementation is provided by Object.

The operation format of the class Strings create a string from arguments
in accord with modifires in the first one.

There is a class Random. An instance of this class could be used to get
pseudo random nambers.

Solve the following issue:
- Print a random number in the range from 5 to 10 inclusive

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>