[04-ADVANCE] add solver of guessed number problem
In Java to run a thread of execution the class Thread MUST BE used. The
operation of an instance of this class will be executed on a dedicated
thread when the operation start will be called.

Create a solver of the guessed number problem that could be run on a
dedicated thread. The solver will be used in the following commits.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>