[06-COLLECTIONS] declare set to solve car park
It could be REQUIRED to object to keep equal elements within a
collection. There is an interface Set to fullfil such wishes in Java.

Declare a set to manage a set of not unique cars with the following
operations: add, remove, size and clear. Provide tests to test that
additions of a car failed if there is an equal car in the set, removal
of an existing car succeeds.

The implementation of the set will be provided on the following
commits.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>