[06-COLLECTIONS] fix access to doubly linked list internals
The keyword protected does not makes a property of an inner class to be
protected, e.g. a field that could be accessed from a subclass. Thus,
fix assess modifiers of fields of an inner class of LinkedCarList to
allow to be accessed in sub-class.

There is no behaviour change.

Fixes: [06-COLLECTIONS] implement doubly linked list of cars
Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>