[06-COLLECTIONS] solve descrease enumeration problem
Let it is required to keep elements of a collection in an order. To
comform this in an efficient way the collection TreeSet is preferable.

All operations on TreeSet have O(log N) time complexity. To place
elements into TreeSet a way to compare elements MUST BE provided.

Solve the follwoing enumeration problem:
- Generate a sequence of one hundred of random numbers in the range from
  0 up to 10 exclusivly.
- Use TreeSet to sort this number in the reverse order.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>