[06-COLLECTIONS] describe interface queue
Beside interfaces Set, List and Map there is an interface Queue in Java.
This interface extends the interface Collection and declares operation
for a First In First Out queue: to put an element at the tail of a queue,
to retrieve and to remove the element at the head of a queue.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>