[07-GENERICS] provide class to keep generic type
It is possible to request the Java translator to try to check that a
variable of a class is assigned with instancies of a specific type,
defined when an instance of this class is declared. The such type is
called a generic type in Java.

Solve the follwoing boxing issue:
- Create a class Box that could keep a variable of any type
- Declare this variable with a generic type.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>