[08-MULTITHREADING] provide one user cashpoint
A block of logic that MUST BE executed by only a one task at a time
is called a critical section. To ensure it, a monitor can be used.

A monitor is an object that could be either acquired or released. If
released, it could be acquired by a single task and others will be
paused. In Java, any instance of a class could be used as monitor.

Solve the following cashpoint problem:
- Create a type and provide an operation that takes a user's name and
  the amount to be withdrawn from her account.
- The operation MUST print that the user has come, wait for 2 seconds,
  then print that the amount is withdrawn or that there is no enough
  money.
- The only one could be served at a time. The operation *join* MUST
  BE used to make other tasks to wait.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>