[08-MULTITHREADING] count up and down atomic
To increment or to decrement an integral value means a sequence of three
operations:
- load the current value of the attribute into a register
- increment or decrement the value
- write the result into the attribute

Some hardware platforms allow to perform such operation at once,
atomically, without the need to use a monitor. For such a case, there
are atomic wrappers in Java.

A variable could be declared with the keyword volatile:
- it prevents rearranging of the instructions and providing a task specific
  copy.
- brings updating of memory and caches on writing

Use an atomic type to solve the counter problem:
- provide a type with the integer attribute *count*
- increment the attribute one thousand times on a one task and decrement
  it the same number of times on an other one
- the result value MUST BE zero

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>