[08-MULTITHREADING] add logic to run jobs on task poll
There are ways to wait for the completion of tasks:
- store references onto task and join on them in a loop
- use CountDownLatch to wait for termination of a specific number of
  tasks

It is not rational to create a large number of tasks at the same time.
Use an ExecutorService to create a poll of a specified number of tasks
and perform jobs on this tasks.

In further commit, this logic will be used to solve the three
computational problem.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>