[08-MULTITHREADING] solve transfer of value problem
Let there is a class with two non negative integer attributes. Provide
two operation, a one to transfer some value from the first attribute to
the second one and an other one to transfer some value from the second
one to the first one. The values of attributes MUST BE kept non
negative. Each attribute MUST BE synchronised with its own monitor.

Provide two tasks, a one to call a one operation and an other one to
call an other one.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>