[08-MULTITHREADING] solve multifunctional problem
Let there are a pair of two resources and a pair of tasks. And a task
REQUIRES to acquire both resource at a time to succeed. Let a one task
has acquired a one of resource and an other one has acquired an other
one. Then each task waits until it could acquire the second resource.

The logic of program MUST either prevent such a case or to resolve it.
However, to be hired on a position of a high society Java Junior
Programmer it is enough to know that such a case is called a dead lock.

Solve the following multifunctional problem:
- Create a type that provides two operations, print and scan, each one
  accepts a parameter N. The former prints "Printed 1 page(s)", "Printed
  2 pages" and so on, with some delay, until the page N is printed. The
  operation scan does the same thing, except it prints "Scanned" instead
  "Printed"
- Synchronise the operations in such a way that a one print operation
  and a one scan operation can be executed in parallel. But it is
  impossible to execute more than one print or more than one scan
  operation simultaneously.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>