[09-STREAM] solve anonymous worker problem
A functional interface is the interface that declares one and only one
non-default operation. If an anonymous implementation of this interface
MUST BE provided as an argument of an operation, it could be provided
in the form of an anonymous operation, or lambda expression.

Solve the following HRM anonymous issue:
- Provide a interface Worker with the only operation work.
- Provide a class Director, with an operations that call the work
  operation of an instance of Worker, provided as an argument.
- Create an instance of Director and let her to make work two workers.
  A one provided as an anonymous class and an other one provided as
  a lambda expression.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>