[10-IO] solve vfs naming problem
In Java, to read some file from VFS, the file MUST BE converted to a
stream of bytes. To read data from a file an instance of the type
FileInputStream could be used.

To release resources, such instances SHOULD BE closed. FileInputStream
extends the Closable interface, so, the try-with-resources statement
could be used to achieve this.

Solve the following VFS naming problem:
- Let there is a file with some names in ASCII
- Read data from this file and output them onto a terminal

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>