[10-IO] solve vfs name space problem
In Java, the type String is immutabe. If it required to modify a string,
use StringBuilder or StringBuffer instead. To improve performance, read
data into an array, a few bytes at once. To read data as a stream of
characters, InputStreamReader could be used.

Solve the following VFS name space problem:
- Create a file with a single line of names, separated by spaces
- Read names into a list of names
- Leave only names that starts with the latter A
- Output the result onto a terminal

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>