[10-IO] solve vfs page problem
There are types to handle data as byte streams: InputStream and
OutputStream, FileInputStream and FileOutputStream, BufferedInputStream
and BufferedOutputStream. There are type to handle data as character
streams: Reader and Writer, FileReader and FileWriter, BufferedReader
and BufferedWriter, InputStreamReader and OutputStreamWriter. And there
is a type RandomAccessFiles to handle files like arrays of bytes.

Solve the following VFS page problem:
- Consider that a page size is 300 bytes
- Write a program that requests a number of a pages within a document
- Then it outputs the requested page onto a terminal and continue to
  request a number again
- Or it terminates if the word exit entered

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>