[11-CORE] describe how java works
Java translators translate a program written in Java into code for the
Java machine, byte-code. Java interpreters, JVMs, perfrom this code on
devices.

A JVM and a Java standard library form a runtime environment, JRE. A set
of a Java translator, JRE and some utilities form a development kit, JDK.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>