[11-CORE] describe interfaces and abstract classes
An interface is an agreement on the behaviour of an object. It defines
what an object can do, not how an object does it. The possibility to
declare default operations and so on has been added to simplify
supporting backward compatibility.

Controversially, abstract classes allows to declare some common logic
that implements behaviour.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>