[11-CORE] describe exceptions and finalise
There are unchecked and checked exceptions. If thrown, an unchecked
exception means a logic error in the program. Checked exception MUST BE
caught in the program.

There is the Error type. If thrown, it means system errors, outside
responsibility of the program.

Types Exception and Error are inherited from the Throwable type.

There is the finalise operation, called before an object is destroyed by
the garbage collector. It is incorrect to rely on this operation due it
is unpredictable when this operation is called.

Signed-off-by: Boris Stankevich <microsoft-wanted@yandex.ru>